gui: presentation: commits: limit pipeset (pre-)rendering
Limit the number of (pre-)rendered pipesets which are visible when
rendering a git graph to the actual currently visible number of commits.
This is important for large repositories with a large number of commits
when scrolling past the first 200 commits, which currently triggers a
load of the complete commit history[1]. If this happens and the git
graph rendering is enabled a potentially large number of pipes (~138k
for the linux kernel) is allocated.

[1]: https://github.com/jesseduffield/lazygit/blob/d56a9cf73e6a34550e42d2627c011c4b73e5a975/pkg/gui/controllers/local_commits_controller.go#L1265

Signed-off-by: Stefan Kerkmann <s.kerkmann@pengutronix.de>